+++ scout/server/blueprints/variants/views.py
@@ -60,6 +60,16 @@ def variants(institute_id, case_name):
     user_obj = store.user(current_user.email)
 
     if request.method == "POST":
+        if request.form.getlist("dismiss"):  # dismiss a list of variants
+            controllers.dismiss_variant_list(
+                store,
+                institute_obj,
+                case_obj,
+                "variant.variant",
+                request.form.getlist("dismiss"),
+                request.form.getlist("dismiss_choices"),
+            )
+
         form = controllers.populate_filters_form(
             store, institute_obj, case_obj, user_obj, category, request.form
         )
@@ -231,6 +241,16 @@ def sv_variants(institute_id, case_name):
     if request.form.get("hpo_clinical_filter"):
         case_obj["hpo_clinical_filter"] = True
 
+    if request.form.getlist("dismiss"):  # dismiss a list of variants
+        controllers.dismiss_variant_list(
+            store,
+            institute_obj,
+            case_obj,
+            "variant.sv_variant",
+            request.form.getlist("dismiss"),
+            request.form.getlist("dismiss_choices"),
+        )
+
     # update status of case if visited for the first time
     controllers.activate_case(store, institute_obj, case_obj, current_user)
     form = controllers.populate_sv_filters_form(store, institute_obj, case_obj, category, request)
@@ -273,6 +293,16 @@ def cancer_variants(institute_id, case_name):
 
     user_obj = store.user(current_user.email)
     if request.method == "POST":
+        if request.form.getlist("dismiss"):  # dismiss a list of variants
+            controllers.dismiss_variant_list(
+                store,
+                institute_obj,
+                case_obj,
+                "variant.variant",
+                request.form.getlist("dismiss"),
+                request.form.getlist("dismiss_choices"),
+            )
+
         form = controllers.populate_filters_form(
             store, institute_obj, case_obj, user_obj, category, request.form
         )