+++ CHANGELOG.md
@@ -11,6 +11,8 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 - Direct use of HPO list for Clinical HPO SNV (and cancer SNV) filtering
 - PDF coverage report redirecting to login page
 ### Changed
+- Remove the option to dismiss single variants from all variants pages
+- Bulk dismiss SNVs, SVs and cancer SNVs from variants pages
 
 
 ## [4.21]